Modularize ReqResp handlers
