+++ packages/lodestar/src/sync/reqResp/reqResp.ts
@@ -2,22 +2,21 @@
  * @module sync
  */
 
-import {GENESIS_SLOT} from "@chainsafe/lodestar-beacon-state-transition";
 import {IBeaconConfig} from "@chainsafe/lodestar-config";
-import {MAX_REQUEST_BLOCKS, phase0} from "@chainsafe/lodestar-types";
+import {phase0} from "@chainsafe/lodestar-types";
 import {ILogger, LodestarError} from "@chainsafe/lodestar-utils";
 import PeerId from "peer-id";
 import {IBeaconChain} from "../../chain";
-import {Method, ReqRespEncoding, RpcResponseStatus} from "../../constants";
+import {Method, ReqRespEncoding} from "../../constants";
 import {IBeaconDb} from "../../db";
 import {createRpcProtocol, INetwork, NetworkEvent} from "../../network";
-import {ResponseError} from "../../network/reqresp/response";
 import {handlePeerMetadataSequence} from "../../network/peers/utils";
 import {syncPeersStatus} from "../utils/sync";
 import {assertPeerRelevance} from "../utils/assertPeerRelevance";
 import {IReqRespHandler} from "./interface";
-import {IBlockFilterOptions} from "../../db/api/beacon/repositories/blockArchive/abstract";
 import {IBeaconMetrics} from "../../metrics";
+import {onBeaconBlocksByRange} from "../../network/reqresp/handlers/beaconBlocksByRange";
+import {onBeaconBlocksByRoot} from "../../network/reqresp/handlers/beaconBlocksByRoot";
 
 export interface IReqRespHandlerModules {
   config: IBeaconConfig;
@@ -113,10 +112,10 @@ export class BeaconReqRespHandler implements IReqRespHandler {
         yield* this.onMetadata();
         break;
       case Method.BeaconBlocksByRange:
-        yield* this.onBeaconBlocksByRange(requestBody as phase0.BeaconBlocksByRangeRequest);
+        yield* onBeaconBlocksByRange(requestBody as phase0.BeaconBlocksByRangeRequest, this.chain, this.db);
         break;
       case Method.BeaconBlocksByRoot:
-        yield* this.onBeaconBlocksByRoot(requestBody as phase0.BeaconBlocksByRootRequest);
+        yield* onBeaconBlocksByRoot(requestBody as phase0.BeaconBlocksByRootRequest, this.db);
         break;
       default:
         throw Error(`Unsupported method ${method}`);
@@ -167,45 +166,6 @@ export class BeaconReqRespHandler implements IReqRespHandler {
     yield this.network.metadata.all;
   }
 
-  private async *onBeaconBlocksByRange(
-    requestBody: phase0.BeaconBlocksByRangeRequest
-  ): AsyncIterable<phase0.SignedBeaconBlock> {
-    if (requestBody.step < 1) {
-      throw new ResponseError(RpcResponseStatus.INVALID_REQUEST, "step < 1");
-    }
-    if (requestBody.count < 1) {
-      throw new ResponseError(RpcResponseStatus.INVALID_REQUEST, "count < 1");
-    }
-    if (requestBody.startSlot < GENESIS_SLOT) {
-      throw new ResponseError(RpcResponseStatus.INVALID_REQUEST, "startSlot < genesis");
-    }
-
-    if (requestBody.count > MAX_REQUEST_BLOCKS) {
-      requestBody.count = MAX_REQUEST_BLOCKS;
-    }
-
-    const archiveBlocksStream = this.db.blockArchive.valuesStream({
-      gte: requestBody.startSlot,
-      lt: requestBody.startSlot + requestBody.count * requestBody.step,
-      step: requestBody.step,
-    } as IBlockFilterOptions);
-    yield* this.injectRecentBlocks(archiveBlocksStream, this.chain, requestBody);
-  }
-
-  private async *onBeaconBlocksByRoot(
-    requestBody: phase0.BeaconBlocksByRootRequest
-  ): AsyncIterable<phase0.SignedBeaconBlock> {
-    const getBlock = this.db.block.get.bind(this.db.block);
-    const getFinalizedBlock = this.db.blockArchive.getByRoot.bind(this.db.blockArchive);
-    for (const blockRoot of requestBody) {
-      const root = blockRoot.valueOf() as Uint8Array;
-      const block = (await getBlock(root)) || (await getFinalizedBlock(root));
-      if (block) {
-        yield block;
-      }
-    }
-  }
-
   private async goodbye(peerId: PeerId, goodbyeCode: GoodByeReasonCode): Promise<void> {
     const reason = GoodbyeReasonCodeDescriptions[goodbyeCode.toString()] || "";
     this.metrics.peerGoodbyeSent.inc({reason});
@@ -224,32 +184,6 @@ export class BeaconReqRespHandler implements IReqRespHandler {
           error: e.message,
         });
         await this.network.disconnect(peerId);
-      }
-    }
-  };
-
-  private injectRecentBlocks = async function* (
-    archiveStream: AsyncIterable<phase0.SignedBeaconBlock>,
-    chain: IBeaconChain,
-    request: phase0.BeaconBlocksByRangeRequest
-  ): AsyncGenerator<phase0.SignedBeaconBlock> {
-    let slot = -1;
-    for await (const archiveBlock of archiveStream) {
-      yield archiveBlock;
-      slot = archiveBlock.message.slot;
-    }
-    slot = slot === -1 ? request.startSlot : slot + request.step;
-    const upperSlot = request.startSlot + request.count * request.step;
-    const slots = [] as number[];
-    while (slot < upperSlot) {
-      slots.push(slot);
-      slot += request.step;
-    }
-
-    const blocks = (await chain.getUnfinalizedBlocksAtSlots(slots)) || [];
-    for (const block of blocks) {
-      if (block) {
-        yield block;
       }
     }
   };
